requirements.txt
feedparser>=6.0
httpx>=0.27
pyyaml>=6.0
jinja2>=3.1
aiosqlite>=0.20
apscheduler>=3.10
loguru>=0.7
click>=8.1
python-dateutil>=2.9
markdownify>=0.13
